+++ include/neuron.h
@@ -12,37 +12,37 @@ class Neuron {
 
         unsigned int get_input_size();
 
-        float get_output();
+        double get_output();
 
-        float get_weighted_sum();
+        double get_weighted_sum();
 
-        float get_weight_value(unsigned int weight);
+        double get_weight_value(unsigned int weight);
 
         void reset_nodeVal();
 
-        float propogate(const std::vector<float>& inputs);
+        double propogate(const std::vector<double>& inputs);
 
-        void update_nodeVal(float nodeVal);
+        void update_nodeVal(double nodeVal);
 
-        void update_weights(unsigned int weight, float input, float learning_rate);
+        void update_weights(unsigned int weight, double input, double learning_rate);
 
-        void update_bias(float learning_rate);
+        void update_bias(double learning_rate);
 
     
     private:
-        float bias_;
-        float weighted_sum_;
-        float output_;
+        double bias_;
+        double weighted_sum_;
+        double output_;
         unsigned int activation_function_;
         //need gradient for weights and biases
-        float nodeVal_;
+        double nodeVal_;
 
         unsigned int input_size_;
-        std::vector<float> weights_;
+        std::vector<double> weights_;
 
-        float activation_function(float input);
+        double activation_function(double input);
 
-        float activation_function_derivative(float input);
+        double activation_function_derivative(double input);
 };
 
 #endif
